Lemoncello/requirements.txt
asgiref==3.2.3
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
backcall==0.1.0
cairocffi==1.0.2
CairoSVG==2.3.1
certifi==2019.9.11
cffi==1.12.3
channels==2.3.1
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
cryptography==2.8
cssselect2==0.2.1
cx-Oracle==7.1.2
daphne==2.3.0
decorator==4.4.1
defusedxml==0.6.0
Django==2.0.7
django-bootstrap4==0.0.8
django-multiselectfield==0.1.8
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
Flask==1.1.1
gunicorn==19.9.0
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipython==7.9.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
MarkupSafe==1.1.1
ndg-httpsclient==0.5.1
numpy==1.17.4
pandas==0.25.3
parso==0.5.1
pdfkit==0.6.1
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.10
pusher==2.1.4
pyasn1==0.4.8
pycparser==2.19
pydotplus==2.0.2
pydub==0.23.1
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.5
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2018.5
redis==2.10.6
reportlab==3.5.21
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.2
six==1.12.0
sklearn==0.0
tinycss2==1.0.2
traitlets==4.3.3
Twisted==19.10.0
txaio==18.8.1
urllib3==1.25.7
wcwidth==0.1.7
WeasyPrint==47
webencodings==0.5.1
Werkzeug==0.16.0
whitenoise==4.1.2
xhtml2pdf==0.2.3
zope.interface==4.7.1